--- brainstat/stats/__init__.py
@@ -1,1 +1,0 @@
-""" The statistics tools of BrainStat """
