+++ brainstat/tests/test_fdr.py
@@ -1,8 +1,9 @@
 import numpy as np
 import pickle
 from .testutil import datadir
-from brainstat.stats.multiple_comparisons import fdr
-
+from brainstat.stats._multiple_comparisons import fdr
+from brainstat.stats.SLM import SLM
+from brainstat.stats.terms import Term
 
 def dummy_test(infile, expfile):
 
@@ -11,47 +12,12 @@ def dummy_test(infile, expfile):
     idic = pickle.load(ifile)
     ifile.close()
 
-    slm = {}
-    slm["t"] = idic["t"]
-    slm["df"] = idic["df"]
-    slm["k"] = idic["k"]
-
-    # check other potential keys in input
-    if "dfs" in idic.keys():
-        slm["dfs"] = idic["dfs"]
-
-    if "resl" in idic.keys():
-        slm["resl"] = idic["resl"]
-
-    if "tri" in idic.keys():
-        slm["tri"] = idic["tri"]
-
-    if "lat" in idic.keys():
-        slm["tri"] = idic["tri"]
-
-    if "mask" in idic.keys():
-        mask = idic["mask"]
-    else:
-        mask = None
-
-    # non-sense input for _fdr, but potentially produced by SurfStatLinMod
-    if "du" in idic.keys():
-        slm["du"] = idic["du"]
-
-    if "c" in idic.keys():
-        slm["c"] = idic["c"]
-
-    if "ef" in idic.keys():
-        slm["ef"] = idic["ef"]
-
-    if "sd" in idic.keys():
-        slm["sd"] = idic["sd"]
-
-    if "SSE" in idic.keys():
-        slm["SSE"] = idic["SSE"]
+    slm = SLM(Term(1), Term(1))
+    for key in idic.keys():
+        setattr(slm, key, idic[key])
 
     # run _fdr
-    outdic = fdr(slm, mask)
+    outdic = fdr(slm)
 
     # load expected outout data
     efile = open(expfile, "br")
